requirements.txt
python-dotenv==0.20.0
requests==2.31.0
Flask==3.0.3
httpx==0.27.0
requests-toolbelt==0.9.1